# Optimize efficiency