--- a/pgdrive/pg_config/body_name.py
+++ b/pgdrive/pg_config/body_name.py
@@ -6,3 +6,4 @@ class BodyName:
     Ego_vehicle = "Ego Vehicle"
     Ego_vehicle_top = "Ego Vehicle Top"
     Traffic_vehicle = "Traffic Vehicle"
+    Lane = "Lane"