--- a/pgdrive/scene_creator/__init__.py
+++ b/pgdrive/scene_creator/__init__.py
@@ -1,1 +1,1 @@
-from pgdrive.scene_creator.basic_utils import Decoration, Goal, get_road_bounding_box, check_lane_on_road
+