--- a/pgdrive/scene_creator/blocks/constants.py
+++ b/pgdrive/scene_creator/blocks/constants.py
@@ -1,5 +1,5 @@
-from pgdrive.scene_creator.lanes.lane import LineType
 from pgdrive.pg_config.collision_group import CollisionGroup
+from pgdrive.scene_creator.lanes.lane import LineType
 
 
 class BlockDefault: