--- a/pgdrive/utils/math_utils.py
+++ b/pgdrive/utils/math_utils.py
@@ -28,7 +28,7 @@ def time_me(fn):
 
 
 def norm(x, y):
-    return math.sqrt(x ** 2 + y ** 2)
+    return math.sqrt(x**2 + y**2)
 
 
 def clip(a, low, high):