@lqy0057 2021-01-08
I improve the efficiency of the environment up to 80%~100%. The only problem is that the new method will be invalid when vehicles get out of road. I **give up** to fall back to ```get_closest_lane_index()``` when the return value of ```ray_localization``` return ```None```. When this happen, the current lane index of all vehicles will be same as last frame.