@lqy0057 2021-01-08
Run ```profile_pgdrive.py```, got these:
![optimize](https://user-images.githubusercontent.com/24489294/103980680-49927000-51bb-11eb-9ae1-179b1887565a.png)
Fig. 1, Optimize branch
![main](https://user-images.githubusercontent.com/24489294/103980705-5616c880-51bb-11eb-86a6-f589e33d15c6.png)
Fig. 2, Main branch
The time used in ```update_states()``` is decreased from 12905ms -> 5067ms, in which ```get_closest_lane()``` is decreased from 8313ms->510ms!!!!!!!!
Although it is amazing,  the time spent in ```doPhysics()``` is much ```higher``` than before, resulting in the total performance of this branch is **lower** than main branch. I am going to optimize it, and then we will improve the efficiency a lot.