@lqy0057 2021-01-08
By assigning collision group, I solved the problem. The efficiency test results are reported by ```python profile_pgdrive.py```

**Test 1**:
Setting: 3 block, 3 traffic vehicle:
Old: 487 fps
New: 600 fps

**Test 2**:
Setting: 5 block, 3 traffic vehicle:
Old: 393 fps
New: 524 fps

**Test 3**:
Setting: 7 block, 4 traffic vehicle:
Old: 311 fps
New: 453 fps

**Test 4**:
Setting: 20 block, 8 traffic vehicle:
Old: 106 fps
New: 216 fps

But I am still not satisfied, and will try some new methods to optimize the ```doPhysics()```