@lqy0057 2021-01-11
![bottle_neck](https://user-images.githubusercontent.com/24489294/104200220-5c2ed280-5463-11eb-9efd-0af3ce48e19a.png)
Time consuming in ```get_closest_lane()```  is a **constant** now, and the last bottleneck is ```neighbour_vehicles()``` .